requirements.txt
Django == 3.2
psycopg2
psycopg2-binary
djangorestframework
django-cors-headers
django-filter
Pillow
django-debug-toolbar
djangorestframework-jwt
django-allauth
django-rest-auth